Refactoring post processing, next stage
